Network learning improvement + added AccuracyEndCondition, CrossValidationDataProvider, DataProvider, EndCondition, EpochsCountEndCondition, ErrorEndCondition, InputsAndTargets, LearningAndValidationData, LearningStatisticsProvider, LearningSupervisor, Statistics, TimeEndCondition # changed type of target values from double to int in ErrorFunction, MultiLayerNeuralNetwork, NeuralNetwork, Neuron, OutputLayerNeuron, OutputNeuronsLayer, SingleLayerNeuralNetwork, SingleLayerNeuron, SingleNeuronsLayer, SumSquaredErrorFunction # renamed toArray method to toDoubleArray in NumericData + added toIntArray method to NumericData

+++ network/src/main/java/com/github/adrian99/neuralnetwork/MultiLayerNeuralNetwork.java
@@ -59,7 +59,7 @@ public class MultiLayerNeuralNetwork extends NeuralNetwork {
     }
 
     @Override
-    protected void calculateNeuronErrors(ErrorFunction errorFunction, double[] targetOutputs) {
+    protected void calculateNeuronErrors(ErrorFunction errorFunction, int[] targetOutputs) {
         outputLayer.calculateNeuronErrors(errorFunction, targetOutputs);
         for (var i = hiddenLayers.length - 1; i >= 0; i--) {
             hiddenLayers[i].calculateNeuronErrors();
